Issue 22. Support Redmine 3.0
changed Redmine Java API dependency from very old (1.19) to 2.2.0.
this required some safe syntax changes plus a big one to support the fact that
all IDs are now immutable in the new version of that library.
as the result, I cannot use issue.setId() in updateTask() method in RedmineTaskSaver anymore.
which is a good thing because it eliminates confusion where ID is loaded from in
    updateTask(String taskId, Issue rmIssue)
- it is from taskId parameter or from rmIssue.getId()?

so, while this is a good API change, this forces some changes in Task Adapter.
we cannot rely on the old "hack" anymore and should delete "taskId" parameter
from
    updateTask(String taskId, Issue rmIssue)
method, which is a part of BasicIssueSaveAPI class.

this won't be a one-line change for JIRA Connector, because it cannot get ID from IssueInput class
- that one is only for creating new JIRA issues, not for updating old ones.
which means we may need to split BasicIssueSaveAPI to
 BasicIssueCreateAPI and BasicIssueUpdateAPI