fixed "update issue" for Redmine 3.0.x introduced when switching from Redmine Java API library 1.19 to 2.2.0.
this is a follow-up to the previous commit. switching to Redmine Java API 2.2.0
required changing the way we update tasks in all connectors.

I deleted taskId parameter from updateTask() method in BasicIssueSaveAPI.
this required adding a wrapper around JIRA's issue because it does not contain an ID
the updater could use.

I moved "check that task can be updated" test from JiraTest to CommonTests and
referenced it from JiraTest class.
also added a new test in RedmineIntegrationTest that calls came "issue updated" test.

side note: I tried adding this to Mantis tests as well, but this revealed a bug in Mantis
 Connector, which is not related to this change, so I will fix it in a separate commit.
I reported it here: https://bitbucket.org/taskadapter/taskadapter/issues/25/once-created-tasks-cannot-be-updated-in