issue #35 unselect "enviroment" field in new JIRA configs
new JIRA versions (like 6.4.11) do not have "environment" field
enabled for tasks by default. let's unselect this field for new configs
to avoid user confusion.

also changed the environment test to use description instead of environment.
AFAIK, the point of the test is to verify that some field is set to default
  when it is empty. this does not seem to be environment-specific.