improved "update" login in GitHub connector
GTaskToGithub was using "remoteId" field, even though other connectors
already switched to gtask.getKey().
"key" is set in the "core" and connectors should just use that one.
the old code worked, but it was not consistent with the other connectors
and it could be broken in the future when remoteId logic is changed.
connectors should only use gTask.getKey() to set the Native Task ID.