fix Linux launcher script generated by Vaadin Application plugin
apparently, Vaadin Application plugin changed format of launcher
script it generates, and my hack to add "export APP_HOME" at the end
of it stopped working because it could not find a tag it was searching.

I updated the tag being searched for. verified that now there is
"export APP_HOME" in the auto-generated 'launcher' script again
(for Linux)