do not attempt to set reporter to empty string
if default value for "reporter" is empty string, GTaskToJira
was setting this to Input Issue builder, and then JIRA was giving
"reporter name must be set" error.
changed the code to not set empty string to reporter name.

updated NewConfigSuggesterTest to account for 1 more field
in mappings list that include JIRA