support anonymous bug reporting via email.
added a setting for anonymous bug reporting to "Configure" page.
if enabled, the app will now submit a bug report to dev@ team
by email with exception details.
this only covers exceptions happened during "export" (load or save)
operation, like class cast exception when converting a field or
connection exception.

added a simple mail sender with a preconfigured "to" and "from" emails.
created a new "to" email in GMail.