redesigned UI - show config panel on a separate page.
the previous version was showing config panel (details)
on the same page as Configs List, in "list - details" element.
now Config details are shown on its own dedicated page,
and Configs List shows a list of links to individual configs,
instead of a combobox with configs.

used the new events model to propagate "config save requested"
and other events instead of passing callbacks around.
the problem with callbacks is that they are expensive to maintain because
they create tight binding between elements.