add app init class TaskAdapterInitListener. restored Google Analytics. event changes.
* TaskAdapterInitListener class is called with the app is initialized. it can listen to
UI objects being created (unlike TALauncher).
* fixed java/scala classes mix problem visible on some local runs - WebAppContext class
was failing to start because it could not find classes in "build/main/java" folder,
because the scala compiler put them in "build/main/scala" folder (which is the workaround
used in this project to allow java and scala code be call each other).
I implemented a fix for local runs - create an empty folder for local dev runs, until
I completedly remove Scala from the project. see TALauncher class for details.
* replaced "ConfigCreateCompleted" event with direct method call.
it was only used for events (stats) reporting.